Did a massive cleanup of bywater's and some of our code to make this plugin more approachable.
Also I'm trying to move all subroutines to a shared library but the plugin system apparently wants code duplication as it doesn't let me. The main module still needs some work because of the insane if-else chaining but it's already an improvement. The email feature still doesn't work. Apparently bywater factored it out so we need to create a new implementation of that feature as the old implementation no longer works.